Add files via upload
In the electronic world, fraud is quite prevalent. Fraudulent transactions can be found in insurance, Internet auctions, benefit claims, and telecommunication
services. When you engage in any online based transactions, there is high probability that you are dealing with a fraudster. Thus, the ability to identify when
any entity you are engaging with online is not playing by the rules is very important.